user register and login TDD, disabled other tests for now

+++ src/main/java/sofe3980/User.java
@@ -28,7 +28,7 @@ public class User {
 
     // Getters
 
-    public long getUserId() {
+    public int getUserId() {
         return userId;
     }
